refactor: move get_accessible_tenant_ids to roles module
Better place and avoiding circular import
