+++ tests/test_helpers/test_tenants.py
@@ -48,58 +48,6 @@ def test_get_all_tenant_ids(base_data):  # pylint: disable=unused-argument
 
 
 @pytest.mark.django_db
-def test_get_accessible_tenant_ids_none(base_data):  # pylint: disable=unused-argument
-    """Verify that get_accessible_tenant_ids returns an empty list when user is None."""
-    result = tenants.get_accessible_tenant_ids(None)
-    assert result == []
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize('user_id, expected', [
-    (1, [1, 2, 3, 7, 8]),
-])
-def test_get_accessible_tenant_ids_super_users(base_data, user_id, expected):  # pylint: disable=unused-argument
-    """Verify get_accessible_tenant_ids function for super users."""
-    user = get_user_model().objects.get(id=user_id)
-    assert user.is_superuser, 'only super users allowed in this test'
-    result = tenants.get_accessible_tenant_ids(user)
-    assert result == expected
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize('user_id, expected', [
-    (2, [1, 2, 3, 7, 8]),
-])
-def test_get_accessible_tenant_ids_staff(base_data, user_id, expected):  # pylint: disable=unused-argument
-    """Verify get_accessible_tenant_ids function for staff users."""
-    user = get_user_model().objects.get(id=user_id)
-    assert user.is_staff, 'only staff users allowed in this test'
-    result = tenants.get_accessible_tenant_ids(user)
-    assert result == expected
-    result = tenants.get_accessible_tenant_ids(user, roles_filter=[])
-    assert result == expected
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize('user_id, expected', [
-    (3, [1]),
-    (4, [1, 2, 7]),
-    (9, [1, 2, 7]),
-    (23, [2, 3, 8]),
-])
-def test_get_accessible_tenant_ids_no_staff_no_sueperuser(
-    base_data, user_id, expected
-):  # pylint: disable=unused-argument
-    """Verify get_accessible_tenant_ids function for users with no staff and no superuser."""
-    user = get_user_model().objects.get(id=user_id)
-    assert not user.is_staff and not user.is_superuser, 'only users with no staff and no superuser allowed in this test'
-    result = tenants.get_accessible_tenant_ids(user)
-    assert result == expected
-    result = tenants.get_accessible_tenant_ids(user, roles_filter=[])
-    assert not result and isinstance(result, list)
-
-
-@pytest.mark.django_db
 def test_get_all_course_org_filter_list(base_data):  # pylint: disable=unused-argument
     """Verify get_all_course_org_filter_list function."""
     result = tenants.get_all_course_org_filter_list()
@@ -170,20 +118,6 @@ def test_get_course_org_filter_list(base_data, tenant_ids, expected):  # pylint:
         with pytest.raises(ValueError) as exc_info:
             tenants.get_course_org_filter_list(tenant_ids)
         assert str(exc_info.value) == f'Invalid tenant IDs: {expected["invalid"]}'
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize('user_id, expected', [
-    (1, [1, 2, 3, 7, 8]),
-    (2, [1, 2, 3, 7, 8]),
-    (3, [1]),
-    (15, []),
-])
-def test_get_accessible_tenant_ids(base_data, user_id, expected):  # pylint: disable=unused-argument
-    """Verify get_accessible_tenant_ids function."""
-    user = get_user_model().objects.get(id=user_id)
-    result = tenants.get_accessible_tenant_ids(user)
-    assert result == expected
 
 
 @pytest.mark.django_db
